aero_rtf_kit: Allow to set force bootloader pin in runtime
Usually I2C and SPI slave and master devices uses the MSB in the
first byte to define if master wants to read(set to 1) or wants to
write(set to 0) to a register in slave.
But as we are using 0x00 to read the FPGA firmware version
lets keep that way to not break anything and lets just use the
read/write bit inverted.

So here 0x01 will read the value of the pin that forces
stay in bootloader and 0x81 will set the value.