aero-rtf: Fix SPI transaction
spi_tx_ready_to_write_rissing_edge was only being set to high by one
clock cycle of in_CLK but tx_byte_buffer was only being set with a
value after the spi_tx_ready_to_write_rissing_edge is set again to
low, causing it to transmit the previous byte.

That is why it was necessary to read 2 time to correcty read the
version and also why the version was being transmitted in the
first byte not on the second.